pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maix"
version = "0.1.0"
description = "MAIX: Modular API Integration eXecutor"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "PyYAML>=6.0.0",
  "pydantic>=2.8.0"
]

[project.optional-dependencies]
cloudwatch = [
  "boto3>=1.34.0"
]
secrets = [
  "boto3>=1.34.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["maix*"]
